Store less data on items client side
